# swagger.json
{
	"components": {
...
		"schemas": {
			"ProfileDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"gender": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"lastOnline": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/UserDto"
					}
				},
				"required": [
					"id",
					"userId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ERole": {
				"enum": [
					"admin",
					"user",
					"guest"
				],
				"type": "string"
			},
			"EPermissions": {
				"enum": [
					"*",
					"user:view",
					"user:manage",
					"contact:view",
					"contact:manage",
					"contact:*",
					"chat:view",
					"chat:manage",
					"chat:*",
					"message:view",
					"message:manage",
					"message:*",
					"push:manage"
				],
				"type": "string"
			},
			"IPermissionDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/EPermissions"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRoleDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/ERole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"phone": {
						"type": "string"
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"profile": {
						"$ref": "#/components/schemas/ProfileDto"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/IRoleDto"
						},
						"type": "array"
					},
					"directPermissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionDto"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"roles",
					"directPermissions",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EProfileStatus": {
				"enum": [
					"online",
					"offline"
				],
				"type": "string"
			},
			"IProfileUpdateRequestDto": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time"
					},
					"gender": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/EProfileStatus"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"EPrivacyLevel": {
				"enum": [
					"everyone",
					"contacts",
					"nobody"
				],
				"type": "string"
			},
			"PrivacySettingsDto": {
				"properties": {
					"showLastOnline": {
						"$ref": "#/components/schemas/EPrivacyLevel"
					},
					"showPhone": {
						"$ref": "#/components/schemas/EPrivacyLevel"
					},
					"showAvatar": {
						"$ref": "#/components/schemas/EPrivacyLevel"
					}
				},
				"required": [
					"showLastOnline",
					"showPhone",
					"showAvatar"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PublicProfileDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/EProfileStatus"
					},
					"lastOnline": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProfileListDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PublicProfileDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"NotificationSettingsDto": {
				"properties": {
					"muteAll": {
						"type": "boolean"
					},
					"soundEnabled": {
						"type": "boolean"
					},
					"showPreview": {
						"type": "boolean"
					}
				},
				"required": [
					"muteAll",
					"soundEnabled",
					"showPreview"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateNotificationSettingsBody": {
				"properties": {
					"muteAll": {
						"type": "boolean"
					},
					"soundEnabled": {
						"type": "boolean"
					},
					"showPreview": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.1",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/profile/my": {
			"get": {
				"operationId": "GetMyProfile",
				"responses": {
					"200": {
						"description": "Пользователь",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileDto"
								}
							}
						}
					}
				},
				"description": "Получить профиль текущего пользователя.\nЭтот эндпоинт позволяет получить данные профиля пользователя, который выполнил запрос.\nИспользуется для получения информации о текущем пользователе, например, его имени, email, и других данных.",
				"summary": "Получение профиля текущего пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/profile/my/update": {
			"patch": {
				"operationId": "UpdateMyProfile",
				"responses": {
					"200": {
						"description": "Обновленный профиль пользователя",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileDto"
								}
							}
						}
					}
				},
				"description": "Обновить профиль текущего пользователя.\nЭтот эндпоинт позволяет пользователю обновить свои данные, такие как имя, email и другие параметры профиля.",
				"summary": "Обновление профиля текущего пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Обновленные данные профиля",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileUpdateRequestDto",
								"description": "Обновленные данные профиля"
							}
						}
					}
				}
			}
		},
		"/api/profile/my/privacy": {
			"get": {
				"operationId": "GetPrivacySettings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PrivacySettingsDto"
								}
							}
						}
					}
				},
				"description": "Получить настройки приватности.",
				"summary": "Настройки приватности",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdatePrivacySettings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PrivacySettingsDto"
								}
							}
						}
					}
				},
				"description": "Обновить настройки приватности.",
				"summary": "Обновление настроек приватности",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"showAvatar": {
										"$ref": "#/components/schemas/EPrivacyLevel"
									},
									"showPhone": {
										"$ref": "#/components/schemas/EPrivacyLevel"
									},
									"showLastOnline": {
										"$ref": "#/components/schemas/EPrivacyLevel"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/profile/my/delete": {
			"delete": {
				"operationId": "DeleteMyProfile",
				"responses": {
					"200": {
						"description": "Сообщение об успешном удалении",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Удалить профиль текущего пользователя.\nЭтот эндпоинт позволяет пользователю удалить свой профиль из системы.",
				"summary": "Удаление профиля текущего пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/profile/all": {
			"get": {
				"operationId": "GetProfiles",
				"responses": {
					"200": {
						"description": "Список всех профилей с информацией о них",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProfileListDto"
								}
							}
						}
					}
				},
				"description": "Получить все профили.\nЭтот эндпоинт позволяет администраторам получить список всех пользователей системы.\nОн поддерживает пагинацию через параметры `offset` и `limit`.",
				"summary": "Получение всех профилей",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"description": "Смещение для пагинации",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Лимит количества возвращаемых профилей",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/profile/{userId}": {
			"get": {
				"operationId": "GetProfileById",
				"responses": {
					"200": {
						"description": "Пользователь по ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PublicProfileDto"
								}
							}
						}
					}
				},
				"description": "Получить профиль по ID.\nЭтот эндпоинт позволяет получить профиль другого пользователя по его ID. Доступен только для администраторов.",
				"summary": "Получение профиля по ID",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, профиль которого нужно получить",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/profile/update/{userId}": {
			"patch": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Обновленный профиль пользователя",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileDto"
								}
							}
						}
					}
				},
				"description": "Обновить профиль другого пользователя.\nЭтот эндпоинт позволяет администраторам обновлять профиль других пользователей.",
				"summary": "Обновление профиля другого пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, профиль которого необходимо обновить",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Данные для обновления профиля",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProfileUpdateRequestDto",
								"description": "Данные для обновления профиля"
							}
						}
					}
				}
			}
		},
		"/api/profile/delete/{userId}": {
			"delete": {
				"operationId": "DeleteProfile",
				"responses": {
					"200": {
						"description": "Сообщение об успешном удалении",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Удалить профиль другого пользователя.\nЭтот эндпоинт позволяет администраторам удалить профиль другого пользователя из системы.",
				"summary": "Удаление профиля другого пользователя",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"jwt": [
							"role:admin"
						]
					}
				],
				"parameters": [
					{
						"description": "ID пользователя, профиль которого необходимо удалить",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/file/{id}": {
			"delete": {
				"operationId": "DeleteFile",
				"responses": {
					"200": {
						"description": "Количество удаленных файлов (обычно 1, если файл существует)",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Удалить файл.\nЭтот эндпоинт позволяет пользователю удалить файл по его ID. Доступ разрешен только пользователю, который загрузил файл, либо администратору.",
				"summary": "Удаление файла",
				"tags": [
					"Files"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID файла, который нужно удалить",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/notification/settings": {
			"get": {
				"operationId": "GetSettings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationSettingsDto"
								}
							}
						}
					}
				},
				"description": "Получить настройки уведомлений текущего пользователя.",
				"summary": "Настройки уведомлений",
				"tags": [
					"Push"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateSettings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationSettingsDto"
								}
							}
						}
					}
				},
				"description": "Обновить настройки уведомлений.",
				"summary": "Обновление настроек уведомлений",
				"tags": [
					"Push"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateNotificationSettingsBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		},
		{
			"url": "http://0.0.0.0:8181",
			"description": "Local dev"
		},
		{
			"url": "http://147.45.245.104:8181",
			"description": "Host"
		}
	]
}